+++ src/utils/hasContent.ts
@@ -12,6 +12,18 @@ function hasDirective(node: AST.VElement, name: string) {
   );
 }
 
+function hasChildWithDirective(node: AST.VElement, name: string): boolean {
+  return node.children.some((child) => {
+    if (child.type !== "VElement") return false;
+
+    if (hasDirective(child, name)) {
+      return true;
+    }
+
+    return hasChildWithDirective(child, name)
+  });
+}
+
 function hasChildImageWithAlt(node: AST.VElement): boolean {
   return node.children.some((child) => {
     if (child.type === "VElement") {
@@ -46,6 +58,8 @@ function hasContent(
     hasAccessibleDirective(node, accessibleDirectives) ||
     hasDirective(node, "text") ||
     hasDirective(node, "html") ||
+    hasChildWithDirective(node, "text") ||
+    hasChildWithDirective(node, "html") ||
     hasChildImageWithAlt(node)
   );
 }